refactor: png를 svg로 변환
